+++ addon/components/legends/wms-wfs-legend.js
@@ -1,10 +1,1 @@
-import WmsLegendComponent from './wms-legend';
-
-/**
-  Component representing map layer's legend for WMS-WFS-layers.
-
-  @class WmsWfsLegendComponent
-  @extends WmsLegendComponent
-*/
-export default WmsLegendComponent.extend({
-});
+export { default } from './wms-legend';