+++ addon/models/new-platform-flexberry-g-i-s-map-layer.js
@@ -29,15 +29,6 @@ let Model = Projection.Model.extend(Offline.ModelMixin, MapLayerMixin, LayerMode
   */
   _leafletObject: null,
 
-  /**
-    Leaflet layer for filters.
-
-    @property _leafletObjectForFilter
-    @type <a href="http://leafletjs.com/reference-1.2.0.html#layer">L.Layer</a>
-    @private
-  */
-  _leafletObjectForFilter: null,
-
   _anyTextChanged: Ember.on('init', Ember.observer('name', 'description', 'keyWords', function() {
     Ember.run.once(this, '_anyTextCompute');
   })),