+++ tests/dummy/config/environment.js
@@ -1,11 +1,22 @@
 /* jshint node: true */
 
 module.exports = function (environment) {
-  var backendUrl = 'https://flexberry-ember-gis.azurewebsites.net';
+  var backendUrl = 'http://bi-vm1.cloudapp.net:12001';
 
   if (environment === 'development-loc') {
     // Use `ember server --environment=development-loc` command for local backend usage.
     backendUrl = 'http://localhost:63138';
+  }
+
+  if (environment === 'mssql-backend') {
+    // Use `ember server --environment=mssql-backend` command for mssql backend usage.
+    backendUrl = 'https://flexberry-ember-gis.azurewebsites.net';
+  }
+
+  if (environment === 'production') {
+    if (process.argv.indexOf('--postfix=-mssql') >= 0) {
+      backendUrl = 'https://flexberry-ember-gis.azurewebsites.net';
+    }
   }
 
   var ENV = {
@@ -55,13 +66,6 @@ module.exports = function (environment) {
         // This let user to continue work without online connection.
         syncDownWhenOnlineEnabled: false,
       },
-
-      components: {
-        flexberryBoundingbox: {
-          // Map id, that must be shown in this component.
-          mapId: '0248624e-ba4a-4fce-a524-48e385d33f41',
-        }
-      }
     }
   };
 
@@ -116,18 +120,24 @@ module.exports = function (environment) {
   // ember build --gh-pages --brunch=<brunch-to-publish-on-gh-pages>.
   if (process.argv.indexOf('--gh-pages') >= 0) {
     var brunch;
+    var postfix = "";
 
-    // Retrieve brunch name from process arguments.
+    // Retrieve brunch name and postfix from process arguments.
     process.argv.forEach(function(value, index) {
       if (value.indexOf('--brunch=') >=0) {
         brunch=value.split('=')[1];
         return;
       }
+
+      if (value.indexOf('--postfix=') >=0) {
+        postfix=value.split('=')[1];
+        return;
+      }
     });
 
     // Change base URL to force paths to application assets be relative.
-    ENV.baseURL = '/' + ENV.repositoryName + '/' + brunch + '/';
-    ENV.locationType = 'none';
+    ENV.baseURL = '/' + ENV.repositoryName + '/' + brunch + postfix + '/';
+    ENV.locationType = 'hash';
   }
 
   return ENV;